Add read-only option to serve_local_nbd
In case a read-only image is served, the client is expected to set the
NBD_READ_ONLY flag from userspace. Do so if requested.

This fixes issues with certain userspace programs making futile attempts
at writing to the device, e.g. mount requiring 'ro' to be set explicitly
instead of auto-detecting.